add status module
