+++ packages/backend/src/Application.ts
@@ -7,6 +7,7 @@ import { createActivityModule } from './modules/activity/ActivityModule'
 import { ApplicationModule } from './modules/ApplicationModule'
 import { createHealthModule } from './modules/health/HealthModule'
 import { createMetricsModule } from './modules/metrics/MetricsModule'
+import { createStatusModule } from './modules/status/StatusModule'
 import { createTvlModule } from './modules/tvl/TvlModule'
 import { createUpdateMonitorModule } from './modules/update-monitor/UpdateMonitorModule'
 import { Database } from './peripherals/database/shared/Database'
@@ -44,6 +45,7 @@ export class Application {
       createTvlModule(config, logger, http, database, clock),
       createActivityModule(config, logger, http, database, clock),
       createUpdateMonitorModule(config, logger, http, database, clock),
+      createStatusModule(config, logger, database, clock),
     ]
 
     const apiServer = new ApiServer(
