+++ packages/backend/src/modules/tvl/TvlModule.ts
@@ -3,17 +3,14 @@ import { providers } from 'ethers'
 
 import { BlocksController } from '../../api/controllers/BlocksController'
 import { DydxController } from '../../api/controllers/DydxController'
-import { StatusController } from '../../api/controllers/status/StatusController'
 import { TvlController } from '../../api/controllers/tvl/TvlController'
 import { createBlocksRouter } from '../../api/routers/BlocksRouter'
 import { createDydxRouter } from '../../api/routers/DydxRouter'
-import { createStatusRouter } from '../../api/routers/StatusRouter'
 import { createTvlRouter } from '../../api/routers/TvlRouter'
 import { Config } from '../../config'
 import { BalanceUpdater } from '../../core/balances/BalanceUpdater'
 import { BlockNumberUpdater } from '../../core/BlockNumberUpdater'
 import { Clock } from '../../core/Clock'
-import { ConfigReader } from '../../core/discovery/config/ConfigReader'
 import { PriceUpdater } from '../../core/PriceUpdater'
 import { ReportUpdater } from '../../core/reports/ReportUpdater'
 import { CoingeckoQueryService } from '../../peripherals/coingecko/CoingeckoQueryService'
@@ -21,7 +18,6 @@ import { AggregateReportRepository } from '../../peripherals/database/AggregateR
 import { BalanceRepository } from '../../peripherals/database/BalanceRepository'
 import { BalanceStatusRepository } from '../../peripherals/database/BalanceStatusRepository'
 import { BlockNumberRepository } from '../../peripherals/database/BlockNumberRepository'
-import { UpdateMonitorRepository } from '../../peripherals/database/discovery/UpdateMonitorRepository'
 import { PriceRepository } from '../../peripherals/database/PriceRepository'
 import { ReportRepository } from '../../peripherals/database/ReportRepository'
 import { ReportStatusRepository } from '../../peripherals/database/ReportStatusRepository'
@@ -44,7 +40,6 @@ export function createTvlModule(
   }
 
   // #region database
-
   const blockNumberRepository = new BlockNumberRepository(database, logger)
   const priceRepository = new PriceRepository(database, logger)
   const balanceRepository = new BalanceRepository(database, logger)
@@ -55,7 +50,6 @@ export function createTvlModule(
   )
   const reportStatusRepository = new ReportStatusRepository(database, logger)
   const balanceStatusRepository = new BalanceStatusRepository(database, logger)
-  const updateMonitorRepository = new UpdateMonitorRepository(database, logger)
 
   // #endregion
   // #region peripherals
@@ -73,7 +67,6 @@ export function createTvlModule(
     http,
     logger,
   )
-  const configReader = new ConfigReader()
 
   // #endregion
   // #region updaters
@@ -88,7 +81,7 @@ export function createTvlModule(
     coingeckoQueryService,
     priceRepository,
     clock,
-    config.tvl.tokens,
+    config.tokens,
     logger,
   )
   const balanceUpdater = new BalanceUpdater(
@@ -120,24 +113,14 @@ export function createTvlModule(
     aggregateReportRepository,
     reportRepository,
     config.projects,
-    config.tvl.tokens,
+    config.tokens,
     logger,
   )
-  const statusController = new StatusController(
-    priceRepository,
-    balanceStatusRepository,
-    reportStatusRepository,
-    updateMonitorRepository,
-    clock,
-    config.tvl.tokens,
-    config.projects,
-    configReader,
-  )
+
   const dydxController = new DydxController(aggregateReportRepository)
 
   const blocksRouter = createBlocksRouter(blocksController)
   const tvlRouter = createTvlRouter(tvlController)
-  const statusRouter = createStatusRouter(statusController)
   const dydxRouter = createDydxRouter(dydxController)
 
   // #endregion
@@ -160,7 +143,7 @@ export function createTvlModule(
   }
 
   return {
-    routers: [blocksRouter, tvlRouter, statusRouter, dydxRouter],
+    routers: [blocksRouter, tvlRouter, dydxRouter],
     start,
   }
 }
